♻️ Ta i bruk felles funksjon for mapping av vedtaksperioder i cucumber tester
