+++ src/test/kotlin/no/nav/tilleggsstonader/sak/vedtak/cucumberUtils/DataTableMapping.kt
@@ -1,0 +1,25 @@
+package no.nav.tilleggsstonader.sak.vedtak.cucumberUtils
+
+import io.cucumber.datatable.DataTable
+import no.nav.tilleggsstonader.sak.cucumber.DomenenøkkelFelles
+import no.nav.tilleggsstonader.sak.cucumber.mapRad
+import no.nav.tilleggsstonader.sak.cucumber.parseValgfriEnum
+import no.nav.tilleggsstonader.sak.cucumber.parseÅrMånedEllerDato
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.BeregningNøkler
+import no.nav.tilleggsstonader.sak.vedtak.domain.Vedtaksperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.AktivitetType
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.MålgruppeType
+import java.util.UUID
+
+fun mapVedtaksperioder(dataTable: DataTable) =
+    dataTable.mapRad { rad ->
+        Vedtaksperiode(
+            id = UUID.randomUUID(),
+            fom = parseÅrMånedEllerDato(DomenenøkkelFelles.FOM, rad).datoEllerFørsteDagenIMåneden(),
+            tom = parseÅrMånedEllerDato(DomenenøkkelFelles.TOM, rad).datoEllerSisteDagenIMåneden(),
+            målgruppe = parseValgfriEnum<MålgruppeType>(BeregningNøkler.MÅLGRUPPE, rad) ?: MålgruppeType.AAP,
+            aktivitet =
+                parseValgfriEnum<AktivitetType>(BeregningNøkler.AKTIVITET, rad)
+                    ?: AktivitetType.TILTAK,
+        )
+    }
